Use a woraround to avoid InputData and InputDataReference collisions
We access the just created InputData node to ensure it is loaded as InputDataImpl instead of InputDataReferenceImpl. If we
do not do this and call InformationRequirement.setRequiredInput the id of our InputData node is associated with an
InputDataReferenceImpl object. This issue seems to be related to CAM-8888 and CAM-8889. However this issue only occurs when
creating a DMN model programmatically using the parser of camunda-dmn-model everything is fine.